Send request to MainActivity to load NowPlaying
The request is sent from the TopTracksFragment to the MainActivity by means of an intent.
MainActivity deals with this intent in handleIntent(). This intent is loaded with extras
(artistName, trackInfo) before being sent. The main activity, upon receipt of this intent
does one of those 2 things depending if the main activity's layout is a single or double
pane. Single pane: NowPlaying is loaded as a full-screen activity using an intent. Double
pane: the NowPlayingFragment is just added to the main activiy's layout without loading
an additional activity (NowPlayingActivity is not loaded).