Move NowPlaying action menu item to activity
Action menu item was in 2 fragments. This was fine in a 1-pane layout as each fragment was loaded
in a different activity. There was only 1 fragment per activity and 1 activity loaded at any given
time. It was a different story in a 2-pane layout as 2 fragments would be loaded in a single
activiy where each fragment had its own menu with the same action menu item. As a result,
2 identical icons was displayed in the menu.